FEATURE: build script to prepack
when `path-browserify` and `url` are `dependencies` eg external then one gets this error in the web console:

Uncaught Error: Dynamic require of "path-browserify" is not supported
    at tailwindcss.worker.js:13:11

so in this WIP they are `devDependencies`

------

those deps where moved to devDependencies to be able to stub the nodebuiltins - which is good for the dev-user but also a bit unfortunate, as they really increase the bundle size.
* "tailwindcss"
* "tailwindcss-language-service"
* "postcss"
* "postcss-selector-parser"

-------

we don't need the stubs in the demos anymore, so they were removed.